LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.bezierCurveTo.nonfinite-expected.txt

2d.path.bezierCurveTo.nonfinite
bezierCurveTo() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS bezierCurveTo() with Infinity/NaN is ignored
